+++ src/main/java/net/wizards/WizardsMod.java
@@ -1,14 +1,13 @@
 package net.wizards;
 
+import net.fabric_extras.structure_pool.api.StructurePoolConfig;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
-import net.minecraft.util.Identifier;
 import net.spell_engine.api.item.ItemConfig;
 import net.spell_engine.api.loot.LootConfig;
 import net.tinyconfig.ConfigManager;
 import net.wizards.config.Default;
 import net.wizards.config.EffectsConfig;
-import net.wizards.config.WorldGenConfig;
 import net.wizards.effect.Effects;
 import net.wizards.item.Armors;
 import net.wizards.item.Group;
@@ -32,8 +31,8 @@ public class WizardsMod {
             .sanitize(true)
             .constrain(LootConfig::constrainValues)
             .build();
-    public static ConfigManager<WorldGenConfig> worldGenConfig = new ConfigManager<WorldGenConfig>
-            ("world_gen", Default.worldGenConfig)
+    public static ConfigManager<StructurePoolConfig> villageConfig = new ConfigManager<>
+            ("villages", Default.villageConfig)
             .builder()
             .setDirectory(ID)
             .sanitize(true)
@@ -54,7 +53,7 @@ public class WizardsMod {
         Weapons.register(itemConfig.value.weapons);
         Armors.register(itemConfig.value.armor_sets);
         itemConfig.save();
-        worldGenConfig.refresh();
+        villageConfig.refresh();
         Effects.register();
         WizardVillagers.register();
     }