+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.2
 
+- Limit Wizard Tower spawns to 1 per village, thanks to the Structure Pool API
 - Fix some default loot table configurations
 - Tweak villager trade offers
 - Update to latest Spell Engine API